refact: extrair listeners de XP/bonus/valores para stat-field-listeners.mjs (actor-sheet.mjs)
